+++ packages/components/src/components/dropdown-item/dropdown-item.tsx
@@ -7,6 +7,8 @@ import { Component, Element, Prop } from '@stencil/core';
 export class DropdownItem {
     @Element() el: HTMLDivElement;
     @Prop() selected: boolean = false;
+    @Prop() disabled: boolean = false;
+    @Prop() excludefromall: boolean = false;
     @Prop() value: string;
     @Prop() text: string;
 }