+++ packages/map-template/releaseTools/webcomponent.js
@@ -32,6 +32,9 @@ MapsIndoorsMap.propTypes = {
     miTransitionLevel: PropTypes.number,
     category: PropTypes.string,
     language: PropTypes.string,
+    searchAllVenues: PropTypes.bool,
+    useKeyboard: PropTypes.bool,
+    hideNonMatches: PropTypes.bool
 };
 
 const WebMapsIndoorsMap = reactToWebComponent(MapsIndoorsMap, React, ReactDOM);