+++ packages/map-template/src/components/Search/Categories/Categories.jsx
@@ -12,6 +12,8 @@ import selectedCategoryState from "../../../atoms/selectedCategoryState";
 import { usePreventSwipe } from '../../../hooks/usePreventSwipe';
 import { useIsDesktop } from "../../../hooks/useIsDesktop";
 import getActiveCategory from "../../../helpers/GetActiveCategory";
+import isBottomSheetLoadedState from "../../../atoms/isBottomSheetLoadedState";
+import categoryState from "../../../atoms/categoryState";
 
 /**
  * Show the categories list.
@@ -44,6 +46,9 @@ function Categories({ onSetSize, getFilteredLocations, searchFieldRef }) {
     const scrollableContentSwipePrevent = usePreventSwipe();
 
     const [activeCategory, setActiveCategory] = useState();
+
+    const isBottomSheetLoaded = useRecoilValue(isBottomSheetLoadedState);
+    const category = useRecoilValue(categoryState);
 
     /**
      * Communicate size change to parent component.
@@ -132,18 +137,19 @@ function Categories({ onSetSize, getFilteredLocations, searchFieldRef }) {
     useEffect(() => {
         if (selectedCategory) {
             getActiveCategory().then(activeCategory => setActiveCategory(activeCategory));
+            setSize(snapPoints.MAX);
         }
     }, [selectedCategory]);
 
     /*
-     * Use the active category state to scroll into view
-     * and center the category.
+     * If the active category is a prop/query parameter and the bottom sheet is loaded, 
+     * then scroll into view and center the active category. 
      */
     useEffect(() => {
-        if (activeCategory) {
-            activeCategory.scrollIntoView({ behavior: 'smooth', block: 'center', inline: 'center' })
+        if (activeCategory && category !== undefined && (isDesktop || (!isDesktop && isBottomSheetLoaded))) {
+            activeCategory.scrollIntoView({ behavior: 'smooth', block: 'nearest', inline: 'center' });
         }
-    }, [activeCategory]);
+    }, [activeCategory, category, isBottomSheetLoaded]);
 
     return (
         <div className="categories prevent-scroll" {...scrollableContentSwipePrevent}>